Refactor IconManager
