+++ lib/settings/cubit/settings_cubit.dart
@@ -4,7 +4,7 @@ import 'dart:ui';
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
-import 'package:nyrna/infrastructure/icon_manager/icon_manager.dart';
+import 'package:nyrna/settings/icon_manager.dart';
 import 'package:nyrna/settings/settings_service.dart';
 import 'package:nyrna/theme/styles.dart';
 import 'package:window_size/window_size.dart' as window;
