+++ lib/settings/icon_manager.dart
